drm/imagination: fix UBSAN shift-out-of-bounds issues
Fix three shift-out-of-bounds issues, e.g.:

  [   90.147149] UBSAN: shift-out-of-bounds in /home/fbinns/dev/img-open-source-graphics/repos/linux-dev/drivers/gpu/drm/imagination/pvr_hwrt.c:167:39
  [   90.160274] shift exponent 32 is too large for 32-bit type 'int'
  [   90.166368] CPU: 1 PID: 352 Comm: deqp-vk Not tainted 6.6.0-rc2-g0bbd46616540-dirty #1
  [   90.166390] Hardware name: Texas Instruments AM625 SK (DT)
  [   90.166398] Call trace:
  [   90.166403]  dump_backtrace+0x94/0xec
  [   90.166427]  show_stack+0x18/0x24
  [   90.166438]  dump_stack_lvl+0x48/0x60
  [   90.166451]  dump_stack+0x18/0x24
  [   90.166460]  ubsan_epilogue+0x10/0x44
  [   90.166473]  __ubsan_handle_shift_out_of_bounds+0x98/0x134
  [   90.166484]  get_cr_multisamplectl_val+0x198/0x454 [powervr]
  [   90.166523]  pvr_hwrt_dataset_create+0x210/0x7c4 [powervr]
  [   90.166544]  pvr_ioctl_create_hwrt_dataset+0x48/0xf0 [powervr]
  [   90.166566]  drm_ioctl_kernel+0xc4/0x1b0
  [   90.166576]  drm_ioctl+0x240/0x550
  [   90.166583]  __arm64_sys_ioctl+0xa8/0xec
  [   90.166594]  invoke_syscall+0x44/0x104
  [   90.166606]  el0_svc_common.constprop.0+0x40/0xe0
  [   90.166616]  do_el0_svc+0x1c/0x28
  [   90.166626]  el0_svc+0x40/0xe8
  [   90.166637]  el0t_64_sync_handler+0xc0/0xc4
  [   90.166647]  el0t_64_sync+0x190/0x194

+++ drivers/gpu/drm/imagination/pvr_hwrt.c
@@ -164,11 +164,11 @@ get_cr_multisamplectl_val(u32 samples, bool y_flip, u64 *value_out)
 		return -EINVAL;
 
 	for (u32 i = 0; i < 8; i++) {
-		value |= sample_positions[idx].x[i] << (i * 8);
+		value |= ((u64)sample_positions[idx].x[i]) << (i * 8);
 		if (y_flip)
-			value |= ((16 - sample_positions[idx].y[i]) & 0xf) << (i * 8 + 4);
+			value |= (((u64)(16 - sample_positions[idx].y[i]) & 0xf)) << (i * 8 + 4);
 		else
-			value |= (sample_positions[idx].y[i]) << (i * 8 + 4);
+			value |= ((u64)sample_positions[idx].y[i]) << (i * 8 + 4);
 	}
 
 	*value_out = value;
